The table size is always 32-bit, but files.size() will return a 64-bit value on 64-bit builds, so I added a cast to silence the warning

+++ src/ArchiveBuilder.cpp
@@ -124,7 +124,7 @@ namespace ZAP
 		writeField(stream, &COMPRESSION); // Compression
 
 		// Build lookup table
-		std::uint32_t tableSize = files.size();
+		std::uint32_t tableSize = static_cast<std::uint32_t>(files.size());
 		writeField(stream, &tableSize); // Table size
 
 		std::uint32_t *fillIn = new std::uint32_t[tableSize];
